Replace cli with plain text due to issue with generating reports using cli-table more than once.
